requirements.txt
asgiref==3.8.1
click==8.2.1
colorama==0.4.6
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
django-redis==5.4.0
django-rq==3.0.1
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
redis==6.2.0
rq==2.3.3
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
coverage==7.8.1
